Move declarations of non-static functions in freetypeui.c to own file
